refactor: project to more modular files
